+++ zarrs/src/array/array_sync_sharded_readable_ext.rs
@@ -3,6 +3,7 @@ use std::{collections::HashMap, sync::Arc};
 use rayon::iter::{IntoParallelIterator, ParallelIterator};
 use rayon_iter_concurrent_limit::iter_concurrent_limit;
 use unsafe_cell_slice::UnsafeCellSlice;
+use zarrs_metadata::ConfigurationSerialize;
 use zarrs_metadata_ext::codec::sharding::ShardingCodecConfiguration;
 use zarrs_storage::byte_range::ByteRange;
 use zarrs_storage::StorageHandle;
@@ -132,7 +133,7 @@ impl ArrayShardedReadableExtCache {
                 .configuration()
                 .expect("valid sharding metadata");
             let sharding_codec_configuration =
-                ShardingCodecConfiguration::try_from(sharding_codec_configuration)
+                ShardingCodecConfiguration::try_from_configuration(sharding_codec_configuration)
                     .expect("valid sharding configuration");
             let sharding_codec = Arc::new(
                 ShardingCodec::new_with_configuration(&sharding_codec_configuration).expect(